Camera and Intro verified
Translated  code to STL from the GUI
Changed the player's entity and different movement

+++ Diablo2/Motor2D/j1Render.cpp
@@ -128,7 +128,8 @@ iPoint j1Render::ScreenToWorld(int x, int y) const
 	return ret;
 }
 
-void j1Render::CenterCamera(float x, float y)
+
+void j1Render::CenterCamera(int x, int y)
 {
 	uint win_x;
 	uint win_y;
